New Shapefile generator, also included code to integrate into new file download capability

+++ src/services/gov/epa/emissions/framework/services/qa/QueryToString.java
@@ -113,7 +113,7 @@ public class QueryToString {
         output.append(colTypes + lineFeeder + colNames + lineFeeder);
     }
     
-    public int getRows(){
+    public long getRows(){
         return rows; 
     }
 }
